Adds one more sanity check for the "BBT edge case"
Also, now that I have DB stuff working again (for the second time) I have found that AGC.dll needs to see the correct view of DSNs when it inits.  This should help those with the latest windows server versions.

git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/FAZR6@953 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/WinTrek/trekinput.cpp
@@ -3057,7 +3057,7 @@ public:
 		//Imago 7/10
 		if (m_bLoad && pmsgBoxLoad == NULL) {
 			ZFile * pzCheck = new ZFile(GetModeler()->GetArtPath() + "/"+INPUTMAP_FILE+"_cloud.mdl",OF_READ);
-			if (pzCheck->GetLength() > 0) { //finally the last sanity check...
+			if (pzCheck && pzCheck->GetLength() > 0) { //finally the last sanity check...
 				delete pzCheck;
 				LoadMap(INPUTMAP_FILE+ZString("_cloud"));
 				DeleteFile(GetModeler()->GetArtPath() + "/"+INPUTMAP_FILE+"_cloud.mdl");
